commit for compile

+++ model/STM32IotNodeI2C.cpp
@@ -32,18 +32,17 @@ namespace codal
 
 int STM32IotNodeI2C::read( uint32_t address, uint8_t reg, uint8_t* data, uint32_t len )
 {
-
  I2C_HandleTypeDef* Handle = ( I2C_HandleTypeDef* ) &_i2c.i2c.handle;
  printf( " \nSTM32IotNodeI2C::read\n" );
- printf( "  Handle          = 0x%08X\n", Handle );
- printf( "  address         = 0x%08X\n", address );
- printf( "  reg             = 0x%02X\n", reg );
- printf( "  len             = 0x%04X\n", len );
- printf( "  data            = 0x%08X\n", data );
+ printf( "  Handle          = 0x%08X\n", ( unsigned int ) Handle );
+ printf( "  address         = 0x%08X\n", ( unsigned int ) address );
+ printf( "  reg             = 0x%02X\n", ( unsigned int ) reg );
+ printf( "  len             = 0x%04X\n", ( unsigned int ) len );
+ printf( "  data            = 0x%08X\n", ( unsigned int ) data );
  if ( HAL_I2C_Mem_Read( Handle, address | 1, ( uint16_t ) reg, I2C_MEMADD_SIZE_8BIT, data, len, 1000 ) == HAL_OK )
  {
   for ( uint32_t i = 0; i < len; i ++ )
-   printf( "   data[ %2d ]     = 0x%02X\n", i, data[ i ] );
+   printf( "   data[ %2d ]     = 0x%02X\n", i, ( unsigned int ) data[ i ] );
   return DEVICE_OK;
  }
  return DEVICE_I2C_ERROR;
@@ -53,13 +52,13 @@ int STM32IotNodeI2C::write( uint32_t address, uint8_t reg, uint8_t* data, uint32
 {
  I2C_HandleTypeDef* Handle = ( I2C_HandleTypeDef* ) &_i2c.i2c.handle;
  printf( " \nSTM32IotNodeI2C::write\n" );
- printf( "  Handle          = 0x%08X\n", Handle );
- printf( "  address         = 0x%08X\n", address );
- printf( "  reg             = 0x%02X\n", reg );
- printf( "  len             = 0x%04X\n", len );
- printf( "  data            = 0x%08X\n", data );
+ printf( "  Handle          = 0x%08X\n", ( unsigned int ) Handle );
+ printf( "  address         = 0x%08X\n", ( unsigned int ) address );
+ printf( "  reg             = 0x%02X\n", ( unsigned int ) reg );
+ printf( "  len             = 0x%04X\n", ( unsigned int ) len );
+ printf( "  data            = 0x%08X\n", ( unsigned int ) data );
  for ( uint32_t i = 0; i < len; i ++ )
-  printf( "   data[ %2d ]     = 0x%02X\n", i, data[ i ] );
+  printf( "   data[ %2d ]     = 0x%02X\n", i, ( unsigned int ) data[ i ] );
  if ( HAL_I2C_Mem_Write( Handle, address, ( uint16_t ) reg, I2C_MEMADD_SIZE_8BIT, data, len, 1000 ) == HAL_OK )
   return DEVICE_OK;
  return DEVICE_I2C_ERROR;
